pyproject.toml
[tool.poetry]
name = "event-server"
version = "0.1.0"
description = "Sync arbitrary events with clients (prototype) "
authors = ["Lalo Martins <lalo.martins@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "event_server", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
tzdata = "^2023.4"
fastapi-jwt-auth = "^0.5.0"
pyotp = "^2.9.0"
pydantic = { extras = ["email"], version = "^2.5.3" }
argon2-cffi = "^23.1.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
bumpversion = "*"
flake8 = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
